ai-service/requirements.txt
fastapi==0.115.0
starlette>=0.37.2,<0.39.0
uvicorn==0.30.6
python-dotenv==1.1.1
google-genai>=1.0.0
langchain-google-genai==2.1.4
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-text-splitters==0.3.11
